control unit tests from system variable

+++ tests/testthat/test-gSD_settings.R
@@ -1,26 +1,27 @@
-context("gSD_settings")
+if(gSD_tests){
+  context("gSD_settings")
 
-test_that("login_CopHub", {
-  expect_is(x <- login_CopHub(username = vars.global$username, password = vars.global$password), "list")
-  expect_true(getOption("gSD.cophub_set"))
-  expect_is(username <- getOption("gSD.cophub_user"), "character")
-  expect_is(password <- getOption("gSD.cophub_pass"), "character")
-})
+  test_that("login_CopHub", {
+    expect_is(x <- login_CopHub(username = vars.global$username, password = vars.global$password), "list")
+    expect_true(getOption("gSD.cophub_set"))
+    expect_is(username <- getOption("gSD.cophub_user"), "character")
+    expect_is(password <- getOption("gSD.cophub_pass"), "character")
+  })
 
-test_that("login_USGS", {
-  expect_is(x <- login_USGS(username = vars.global$username, password = vars.global$password), "list")
-  expect_true(getOption("gSD.usgs_set"))
-  expect_is(username <- getOption("gSD.usgs_user"), "character")
-  expect_is(password <- getOption("gSD.usgs_pass"), "character")
-})
+  test_that("login_USGS", {
+    expect_is(x <- login_USGS(username = vars.global$username, password = vars.global$password), "list")
+    expect_true(getOption("gSD.usgs_set"))
+    expect_is(username <- getOption("gSD.usgs_user"), "character")
+    expect_is(password <- getOption("gSD.usgs_pass"), "character")
+  })
 
-test_that("set_archive", {
-  expect_is(x <- set_archive(vars.global$dir.arc), "list")
-  expect_true(getOption("gSD.archive_set"))
-  expect_is(getOption("gSD.archive"), "character")
-})
+  test_that("set_archive", {
+    expect_is(x <- set_archive(vars.global$dir.arc), "list")
+    expect_true(getOption("gSD.archive_set"))
+    expect_is(getOption("gSD.archive"), "character")
+  })
 
-unlink(vars.global$dir.arc)
+  unlink(vars.global$dir.arc)
 
-## add set_aoi tests
-
+  ## add set_aoi tests
+}
